Adding comments, checking watershed file to make sure it matches initialization time, only updating latest if within 24 h

+++ init_AR_sens.py
@@ -2,6 +2,7 @@ import os, sys, shutil
 import glob
 import datetime as dt
 import numpy as np
+import xarray as xr
 import argparse
 import subprocess
 import configparser
@@ -16,26 +17,31 @@ def init_AR_sens(init, paramfile):
 
   init_dt = dt.datetime.strptime(init, '%Y%m%d%H')
 
-  if (not os.path.exists('{0}/pcp/{1}_auto'.format(conf['metric']['precip_metric_loc'],init))) and \
-      os.path.exists('{0}/pcp/yyyymmddhh_auto'.format(conf['metric']['precip_metric_loc'])):
-    shutil.copy('{0}/pcp/yyyymmddhh_auto'.format(conf['metric']['precip_metric_loc']),'{0}/pcp/{1}_auto'.format(conf['metric']['precip_metric_loc'],init))
+  if (not os.path.exists('{0}/{1}_auto'.format(conf['metric'].get('precip_metric_loc','.'),init))) and \
+      os.path.exists('{0}/yyyymmddhh_auto'.format(conf['metric'].get('precip_metric_loc','.'))):
+    shutil.copy('{0}/yyyymmddhh_auto'.format(conf['metric']['precip_metric_loc']),'{0}/{1}_auto'.format(conf['metric']['precip_metric_loc'],init))
 
-  if (not os.path.exists('{0}/rbasin/{1}_auto'.format(conf['metric']['basin_metric_loc'],init))) and \
-      os.path.exists('{0}/rbasin/yyyymmddhh_auto'.format(conf['metric']['basin_metric_loc'])):
-    shutil.copy('{0}/rbasin/yyyymmddhh_auto'.format(conf['metric']['basin_metric_loc']),'{0}/rbasin/{1}_auto'.format(conf['metric']['basin_metric_loc'],init))
+  if (not os.path.exists('{0}/{1}_auto'.format(conf['metric'].get('basin_metric_loc','.'),init))) and \
+      os.path.exists('{0}/yyyymmddhh_auto'.format(conf['metric'].get('basin_metric_loc','.'))):
+    shutil.copy('{0}/yyyymmddhh_auto'.format(conf['metric']['basin_metric_loc']),'{0}/{1}_auto'.format(conf['metric']['basin_metric_loc'],init))
 
-  if (not os.path.exists('{0}/ivtl/{1}_auto'.format(conf['metric']['ivt_land_metric_loc'],init))) and \
-      os.path.exists('{0}/ivtl/yyyymmddhh_auto'.format(conf['metric']['ivt_land_metric_loc'])):
-    shutil.copy('{0}/ivtl/yyyymmddhh_auto'.format(conf['metric']['ivt_land_metric_loc']),'{0}/ivtl/{1}_auto'.format(conf['metric']['ivt_land_metric_loc'],init))
+  if (not os.path.exists('{0}/{1}_auto'.format(conf['metric'].get('ivt_land_metric_loc','.'),init))) and \
+      os.path.exists('{0}/yyyymmddhh_auto'.format(conf['metric'].get('ivt_land_metric_loc','.'))):
+    shutil.copy('{0}/yyyymmddhh_auto'.format(conf['metric']['ivt_land_metric_loc']),'{0}/{1}_auto'.format(conf['metric']['ivt_land_metric_loc'],init))
 
   conf['locations']['work_dir'] = '{0}/{1}'.format(conf['locations']['work_dir'],init)
   os.makedirs(conf['locations']['work_dir'], exist_ok = True)
   os.chdir(conf['locations']['work_dir'])
 
-  if not os.path.exists('{0}/watershed_precip.nc'.format(conf['locations']['work_dir'])):
+  #  Download the watershed precipitation files.  Remove if this file does not match the initialization time
+  if (not os.path.exists('{0}/watershed_precip.nc'.format(conf['locations']['work_dir']))) and 'basin_metric_loc' in conf['metric']:
     urllib.request.urlretrieve('https://cw3e.ucsd.edu/Projects/QPF/data/watershed_HUC8_eps.nc', 'watershed_precip.nc')
+    ds = xr.open_dataset('watershed_precip.nc', decode_times=False).rename({'time': 'hour'})
+    if ds.attrs['init'] != init:
+      os.remove('watershed_precip.nc')
 
-  if (not os.path.exists('{0}/arr_buoys.txt')) and dt.datetime.now().hour > 0:
+  #  Download the AR Recon buoy files
+  if (not os.path.exists('{0}/arr_buoys.txt')) and dt.datetime.now().hour > 12:
     fout = open('arr_buoys.txt', 'w')
     for yyyy in [2020, 2021, 2022, 2023, 2024, 2025, 2026]:
       urllib.request.urlretrieve('https://cw3e.ucsd.edu/images/CW3E_Obs/DriftingBuoys/ARRecon_{0}_SVP-B_BuoyLocations_latest_SLP.txt'.format(yyyy), 'buoy.txt')
@@ -45,27 +51,33 @@ def init_AR_sens(init, paramfile):
 
     fout.close()
 
-  if (not os.path.exists('{0}/other_buoys.txt')) and dt.datetime.now().hour > 0:
+  #  Download the other drifting buoys
+  if (not os.path.exists('{0}/other_buoys.txt')) and dt.datetime.now().hour > 12:
     urllib.request.urlretrieve('https://cw3e.ucsd.edu/images/CW3E_Obs/DriftingBuoys/BuoyLocations_latest_SLP.txt', 'other_buoys.txt')
 
+  #  Actually run the sensitivity code.  Using system call to make sure there is no overwrite of current variables
   os.chdir('/home11/staff/torn/ens-sens/AR-sens')
   os.system('python run_AR_sens.py --init {0} --param {1}'.format(init,paramfile))
 
+  #  Remove the source Grib files if they are being obtained from AWS or Google Cloud
   if conf['model']['io_module'] == 'ecmwf_aws_down' or conf['model']['io_module'] == 'gefs_aws_down': 
     for rmfile in glob.glob('{0}/f*grb2*'.format(conf['locations']['work_dir'])):
       os.remove(rmfile)
 
+  #  Create a list of metrics for this particular initialization time
   if os.path.exists('{0}/metric_list'.format(conf['locations']['work_dir'])):
     with open('{0}/metric_list'.format(conf['locations']['work_dir']), 'r') as f:
       metlist = [line.rstrip('\n') for line in f]
   else:
     metlist = []
 
+  #  Create html panel for each initialization time for metric, update sidebar and metric list
   ar_sens_html(init, metlist, paramfile)
 
+  #  Place the most recent version of different metrics into special directory for Google Earth.  Only done if within 24 h of initialization
   for mettype in ['pcp1', 'pcp2', 'ivtland1', 'ivtland2', 'ivt1', 'slp1', 'hght1', 'pv1', 'snow1']:
 
-    if os.path.exists('{0}/latest/{1}'.format(conf['locations']['figure_dir'],mettype)):
+    if os.path.exists('{0}/latest/{1}'.format(conf['locations']['figure_dir'],mettype)) and (dt.datetime.now()-init_dt).total_seconds() < 86400.:
 
       for filesuf in ['.png', '.nc', '.tar']:
          for rmfile in glob.glob('{0}/latest/{1}/*{2}'.format(conf['locations']['figure_dir'],mettype,filesuf)):
@@ -74,7 +86,8 @@ def init_AR_sens(init, paramfile):
       if any(mettype in item for item in metlist):
 
         metname = [item for item in metlist if mettype in item][0]
-        
+
+        #  Copy each of the requested files, including metric figure, metric netCDF, tar file of sensitivity grids and sensitivity figures
         shutil.copy('{0}/{1}/{2}/metric.png'.format(conf['locations']['figure_dir'],init,metname),'{0}/latest/{1}/.'.format(conf['locations']['figure_dir'],mettype))
         shutil.copy('{0}/{1}_{2}.nc'.format(conf['locations']['work_dir'],init,metname),'{0}/latest/{1}/metric.nc'.format(conf['locations']['figure_dir'],mettype))
         shutil.copy('{0}/{1}/{1}_{2}_esens.tar'.format(conf['locations']['figure_dir'],init,metname),'{0}/latest/{1}/esens.tar'.format(conf['locations']['figure_dir'],mettype))
@@ -83,8 +96,6 @@ def init_AR_sens(init, paramfile):
           for fhrt in ['048', '072']:
             shutil.copy('{0}/{1}/{2}/sens/{3}/{1}_f{4}_{3}_sens.png'.format(conf['locations']['figure_dir'],init,metname,field,fhrt), \
                         '{0}/latest/{1}/f{2}_{3}_sens.png'.format(conf['locations']['figure_dir'],mettype,fhrt,field))
-
-
 
 
 if __name__ == '__main__':
